Implemented DontDeleteVideo after conversion option.

+++ FairyFileRenamerProject/FairyFileRenamerProject/Mp4ToMp3Converter.cs
@@ -19,15 +19,17 @@ namespace FairyFileRenamerProject
 
         string pathToffmpeg;
 
+        bool deleteVideoAfter = false;
+
         BackgroundWorker worker = new BackgroundWorker();
         
 
-        public Mp4ToMp3Converter(string _inputFile)
+        public Mp4ToMp3Converter(string _inputFile,bool _deleteVideoAfter)
         {
             inputFile = _inputFile;
             outputFile = inputFile.Substring(0, inputFile.Length - 1) + "3";
 
-            
+            deleteVideoAfter = _deleteVideoAfter;
 
             pathToffmpeg = Environment.CurrentDirectory +"\\ffmpeg.exe";
 
@@ -42,8 +44,8 @@ namespace FairyFileRenamerProject
 
             FFMPEG ffmpeg = new FFMPEG(pathToffmpeg);
             string result2 = ffmpeg.RunCommand("-i \"" + inputFile + "\" -vn -f mp3 -ab 192k \"" + outputFile + "\"");
-
-            File.Delete(inputFile);
+            if(deleteVideoAfter)
+                File.Delete(inputFile);
             
         }
         private void worker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
